% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelic_heterogenity_boost.R
\name{allelic_het_simulator}
\alias{allelic_het_simulator}
\title{Allelic heterogeneity simulation and calculation of IVW uplift}
\usage{
allelic_het_simulator(
  betas,
  n1 = 392814,
  n2 = 260405,
  maf_range = c(1e-04, 5e-04, 0.001, 0.0025, 0.005, 0.01),
  enrichment_value = c(1, 5, 10, 20, 30, 50),
  disease_prevalence = 0.005,
  disease_prevalence_scale_stdy2 = 1.5,
  n_simulated_datasets = 1000,
  effect_direction = "increasing",
  meta_compare = FALSE,
  quantile_range = c(0.1, 0.5, 0.9)
)
}
\arguments{
\item{betas}{data.frame whose columns denote the regression coefficients from study 1 (UKB) and separately study 2 (FinGen) studies.}

\item{n1}{numeric variable denoting the sample size of study 1. Default set to UKB sample size.}

\item{n2}{numeric variable denoting the sample size of study 2. Default set to FG sample size.}

\item{maf_range}{numeric vector containing MAF values for study 1. Default considers a range of MAFs starting at the very rare MAF 0.01% and sequentially increasing to moderate rare 1%.}

\item{enrichment_value}{numeric vector denoting the allelic enrichment of study 2. Default considers a range of allelic enrichment values starting at no enrichment (i.e, a value of 1) to 50-fold enrichment.}

\item{disease_prevalence}{numeric vector denoting disease prevalence in study 1. Default is 0.5% which is the same order of mangnitude as the median disease prevalence across the 975 diseases in UKB.}

\item{disease_prevalence_scale_stdy2}{numeric scalar denoting the scaled increase/decrease of disease prevalence in study 2. Default is 1.5, i.e,. 0.75%.}

\item{n_simulated_datasets}{numeric scalar denoting the number of datasets to simulate for each value of MAF and enrichment value.}

\item{effect_direction}{character denoting the direction of association (i.e., risk increasing or decreasing). Default is risk increasing}

\item{meta_compare}{logical denoting whether we perform the primary meta-analysis (i.e., the meta-analysis which combines 2 studies from different populations) or to compare the relative effects of two meta-analyses: (i) 2 studies from different populations and (ii) 2 studies from the same population. Default is FALSE, i.e, to perform our primary meta-analysis.}

\item{quantile_range}{numeric vector denoting the set of quantile summary values to record. Default is {0.1,0.5,0.9}, i.e., the 1st, 5th (median) and 9th deciles.}
}
\value{
Returned are user defined quantile range of simulated and theoretically predicted values of IVW-uplift generated from "n_simulated_datasets" per combination of MAF and enrichment value.
}
\description{
Simulation protocol to generate and estimate IVW-uplift owing to 
allelic-heterogeneity between 2 population cohorts.
}
