% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betas.R
\docType{data}
\name{betas}
\alias{betas}
\title{Genetic association data from UKB and FG}
\format{
A data frame with 467 rows and 2 columns:
\describe{
  \item{ukb}{UKB betas}
  \item{fg}{FG betas}
}
}
\source{
\url{https://www.medrxiv.org/content/10.1101/2021.10.14.21265023v1.article-info}
}
\usage{
betas
}
\description{
A dataset containing regression estimates (i.e., betas) of the genetic associations in UKB and FG.
}
\keyword{datasets}
