% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_clust_code.R
\name{figure_4_mr_clust}
\alias{figure_4_mr_clust}
\title{MR-Clust mixture model fitting}
\usage{
figure_4_mr_clust(figure_4_panel = "a")
}
\arguments{
\item{figure_4_panel}{Figure 4 panel to replicate - default is figure panel "a", i.e., Figure 4a.}
}
\value{
Returned are: estimates of the putative number of clusters in the
sample, complete with allocation probabilities and summaries of the
association estimates for each variant; plots which visualise the allocation
of variants to clusters and; several summaries of the fitting process, i.e.
the BIC and likelihood estimates.
}
\description{
Assessment of clustered heterogeneity in Mendelian randomization analyses
using expectation-maximisation (EM) based model fitting of the MR-Clust
mixture model. Function output includes both data-tables and a visualisation
of the assingment of variants to clusters.
}
