% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflate_factor.R
\name{inflate_factor}
\alias{inflate_factor}
\title{Inflation factor}
\usage{
inflate_factor(sig_z, se_ratio)
}
\arguments{
\item{sig_z}{}

\item{se_ratio}{}
}
\value{
Returned is denominator of IVW-uplift.
}
\description{
Inflation factor
}
