% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_run.R
\name{logit_run}
\alias{logit_run}
\title{Simulated disease endpoints}
\usage{
logit_run(n, maf, pi_y, bta, full_analysis = T)
}
\arguments{
\item{n}{}

\item{maf}{}

\item{pi_y}{}

\item{bta}{}
}
\value{
Return simulated disease endpoint.
}
\description{
Simulated disease endpoints
}
