% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_sums.R
\name{quantile_sums}
\alias{quantile_sums}
\title{Quantile computation}
\usage{
quantile_sums(data, values = c(0.1, 0.5, 0.9))
}
\arguments{
\item{data}{list taken from allelic heterogeneity simulation.}

\item{values}{numeric vector denoting the set of quantile summary values to record. Default is {0.1,0.5,0.9}, i.e., the 1st, 5th (median) and 9th deciles.}
}
\value{
Returned are user defined quantile range of simulated and theoretically predicted values of IVW-uplift generated from "n_simulated_datasets" per combination of MAF and enrichment value.
}
\description{
Quantile computation
}
