% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_sig_z.R
\name{ratio_sig_z}
\alias{ratio_sig_z}
\title{IVW z-score}
\usage{
ratio_sig_z(z1, se1, z2, se2)
}
\arguments{
\item{z1}{z-score study 1}

\item{se1}{standard error study 1}

\item{z2}{a-score study 2}

\item{se2}{standard error study 2}
}
\value{
IVW z-score.
}
\description{
IVW z-score
}
