% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_plot_fun.R
\name{surface_plots}
\alias{surface_plots}
\title{Surface plots}
\usage{
surface_plots(data, meta_compare = FALSE)
}
\arguments{
\item{data}{data.frame computed from "allelic_heterogenetity_boost.R"}

\item{meta_compare}{logical denoting whether we perform the primary meta-analysis (i.e., the meta-analysis which combines 2 studies from different populations) or to compare the relative effects of two meta-analyses: (i) 2 studies from different populations and (ii) 2 studies from the same population. Default is FALSE, i.e, to perform our primary meta-analysis.}
}
\value{
Returned are surface plots of the observed (i.e., simulated) and expected (i.e., theoretically predicted) values for IVW-uplift.
}
\description{
Surface plots
}
