% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_ratio.R
\name{var_ratio}
\alias{var_ratio}
\title{Ratio of variance of binomial RVs}
\usage{
var_ratio(p1, p2)
}
\arguments{
\item{p1}{maf study 1}

\item{p2}{maf study 2}
}
\value{
Returned is denominator of IVW-uplift.
}
\description{
Ratio of variance of binomial RVs
}
