% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_ivw.R
\name{z_ivw}
\alias{z_ivw}
\title{IVW z-score}
\usage{
z_ivw(est1, se1, est2, se2)
}
\arguments{
\item{est1}{regression estimate study 1}

\item{se1}{regression std error study 1}

\item{est2}{regression estimate study 2}

\item{se2}{regression std error study 2}
}
\value{
IVW z-score.
}
\description{
IVW z-score
}
