% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_ratio.R
\name{exp_ratio}
\alias{exp_ratio}
\title{Expected value of standard error ratios}
\usage{
exp_ratio(p1_ast, n1_ast, pi1_cond, p2_ast, n2_ast, pi2_cond)
}
\arguments{
\item{pi1_cond}{}

\item{p1}{}

\item{n1}{}
}
\value{
Return expected value of standard error ratios
}
\description{
Expected value of standard error ratios
}
