% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_run.R
\name{logit_run}
\alias{logit_run}
\title{Simulated disease endpoints}
\usage{
logit_run(n, maf, pi_y, bta)
}
\arguments{
\item{n}{sample size}

\item{maf}{minor allele frequency}

\item{pi_y}{disease prevalence}

\item{bta}{genetic effect size}
}
\value{
Return simulated disease endpoint.
}
\description{
Simulated disease endpoints
}
