% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_sig_z.R
\name{ratio_sig_z}
\alias{ratio_sig_z}
\title{IVW z-score}
\usage{
ratio_sig_z(z1, se1, z2, se2)
}
\arguments{
\item{se1}{}

\item{se2}{}

\item{est1}{}

\item{est2}{}
}
\value{
IVW z-score.
}
\description{
IVW z-score
}
