% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_ratio_exp.R
\name{se_ratio_exp}
\alias{se_ratio_exp}
\title{Expected value of standard error ratios}
\usage{
se_ratio_exp(p1, n1, pi1_cond, p2, n2, pi2_cond)
}
\arguments{
\item{p1}{}

\item{n1}{}

\item{pi1_cond}{}
}
\value{
Return expected value of standard error ratios
}
\description{
Expected value of standard error ratios
}
