% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_plot_fun.R
\name{surface_plots}
\alias{surface_plots}
\title{Surface plots}
\usage{
surface_plots(results)
}
\arguments{
\item{results}{data.frame computed from "allelic_heterogenetity_boost.R"}
}
\value{
Returned are surface plots of the observed (i.e., simulated) and expected (i.e., theoretically predicted) values for IVW-uplift.
}
\description{
Surface plots
}
